Refactor and enhance cryptographic components
- Corrected spelling of "Shannon" in KeyManager and updated related methods.
- Removed unnecessary using directives in KeyTypes.
- Updated PgpBuilder to use consistent naming for PEM separators.
- Improved memory security in Encryption class by ensuring AES keys are cleared after use.
- Added default memory size constant for Argon2 in KeyDerivationBuilder.
- Introduced ArgumentHelper for argument validation across frameworks.
- Added BinaryHelpers for little-endian read/write operations.
- Enhanced CryptographicOperations with constant-time equality checks.
- Implemented HashAlgorithmHelper for hash algorithm management.
- Improved SecureMemoryOperations for constant-time comparisons.
- Added SecureRandomNumberGenerator with an entropy pool for enhanced randomness.
- Introduced Ed25519 signature tests to validate RFC 8032 compliance and signature properties.
- Cleaned up unused references in test files and ensured consistent naming conventions.

+++ src/HeroCrypt/Cryptography/Primitives/Kdf/ScryptCore.cs
@@ -155,8 +155,8 @@ internal static class ScryptCore
             throw new ArgumentException("Output length must be positive", nameof(outputLength));
         }
 
-        // Check memory requirements
-        var memoryRequired = (long)128 * r * n * p;
+        // Check memory requirements (use 128L to ensure long arithmetic and prevent overflow)
+        var memoryRequired = 128L * r * n * p;
         if (memoryRequired > DEFAULT_MAX_MEMORY)
         {
             throw new ArgumentException($"Parameters require too much memory: {memoryRequired} bytes (max: {DEFAULT_MAX_MEMORY})", nameof(n));
